pyproject.toml
[project]
name = "infinityhackthon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-restframework>=0.0.1",
    "djangorestframework-simplejwt>=5.5.0",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "python-dotenv>=1.1.1",
]
